Move the fakeNow function to a utility file and use also on stats page.
