+++ e2e/tests/feature-page.spec.ts
@@ -14,39 +14,8 @@
  * limitations under the License.
  */
 
-import {test, expect, Page} from '@playwright/test';
-
-async function setupFakeNow(page: Page): Promise<void> {
-  // Get fakeNow from UTC to extract the timeZone offset used in the test
-  const fakeNowDateTime = 'Dec 1 2020 12:34:56';
-  const fakeNowFromUTC = new Date(fakeNowDateTime);
-  const offset = fakeNowFromUTC.getTimezoneOffset();
-  const offsetSign = offset < 0 ? '-' : '+';
-  const offsetHours = `${Math.abs(Math.floor(offset / 60))}`.padStart(2, '0');
-  const offsetMinutes = `${Math.abs(offset % 60)}`.padStart(2, '0');
-  const offsetText = `${offsetSign}${offsetHours}:${offsetMinutes}`;
-
-  // Get fakeNow from the test timeZone
-  const fakeNow = new Date(`${fakeNowDateTime}Z${offsetText}`).valueOf();
-
-  // Update the Date accordingly in your test pages
-  await page.addInitScript(`{
-    // Extend Date constructor to default to fakeNow
-    Date = class extends Date {
-      constructor(...args) {
-        if (args.length === 0) {
-          super(${fakeNow});
-        } else {
-          super(...args);
-        }
-      }
-    }
-    // Override Date.now() to start from fakeNow
-    const __DateNowOffset = ${fakeNow} - Date.now();
-    const __DateNow = Date.now;
-    Date.now = () => __DateNow() + __DateNowOffset;
-  }`);
-}
+import { test, expect } from '@playwright/test';
+import { setupFakeNow } from './utils';
 
 test.beforeEach(async ({page}) => {
   await setupFakeNow(page);
